requirements.txt
#
# 本文件列出 csms 服务所需 Python 依赖。
# 包括 FastAPI、Uvicorn、websockets、redis、pydantic、数据库等。
#
fastapi==0.115.2
uvicorn[standard]==0.30.6
websockets==12.0
redis==5.0.8
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
# 数据库
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
# 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
# 日志和监控
python-json-logger==2.0.7
prometheus-client==0.20.0
# HTTP客户端
httpx==0.27.2
# MQTT客户端
paho-mqtt==2.1.0
